+++ deconz/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 8.5.2
 
-- Add configurable OTA update toggles for Bosch, IKEA and LEDVANCE
+- Add configurable OTA update switches for Bosch, IKEA and LEDVANCE (default is off)
 - Add Bosch Smart Home OTA update script (Official repo)
 
 ## 8.5.1